separated types and config for docs
