+++ libs/docs/src/config.ts
@@ -1,0 +1,7 @@
+import { DocVersions } from './types';
+
+export const docVersions = ['2.0.0', '2.1.0', '2.2.0-alpha'] as const;
+
+export const defaultVersion: DocVersions = '2.1.0';
+
+export const githubDocsUrl = 'https://github.com/devfile/devfile-web/tree/main/libs/docs/src/docs';
