+++ libs/docs/src/index.ts
@@ -1,2 +1,2 @@
-export * from './scripts';
 export * from './types';
+export * from './config';
