+++ libs/docs/src/types/index.ts
@@ -1,8 +1,4 @@
-export const docVersions = ['2.0.0', '2.1.0', '2.2.0-alpha'] as const;
-
-export const defaultVersion: DocVersions = '2.1.0';
-
-export const githubDocsUrl = 'https://github.com/devfile/devfile-web/tree/main/libs/docs/src/docs';
+import { docVersions } from '../config';
 
 export type DocVersions = typeof docVersions[number];
 export interface Page {
